TvSetting: soft Ap [1/1]
PD#OTT-65288
PD#SWPL-185644

Problem:
Too long characters will cause failure in starting softap,
so the length needs to be limited. The maximum length of
SSID is 64. The maximum length of password is 64 and the
minimum length is 8.

Solution:
Too long characters will cause failure in starting softap,
so the length needs to be limited. The maximum length of
SSID is 64. The maximum length of password is 64 and the
minimum length is 8.

Verify:
androidU

Change-Id: I8001a6f1e7b679b818fd73213840d203a258bae1
Signed-off-by: junlu.jiang <junlu.jiang@amlogic.com>